Serialization functions in .cpp
