+++ GameEngine/GameEngine/GameEngine.vcxproj
@@ -240,6 +240,7 @@
     <ClCompile Include="Engine\Geometry\MaterialData.cpp" />
     <ClCompile Include="Engine\Geometry\Mesh.cpp" />
     <ClCompile Include="Engine\Physic\BulletCollisionManager.cpp" />
+    <ClCompile Include="Engine\Physic\DynamicMotionState.cpp" />
     <ClCompile Include="Engine\Render\Buffer.cpp" />
     <ClCompile Include="Engine\Render\Data.cpp" />
     <ClCompile Include="Engine\Render\Framebuffer.cpp" />
@@ -260,6 +261,7 @@
     <ClCompile Include="Engine\Render\VertexArray.cpp" />
     <ClCompile Include="Engine\Skinning\AnimationChannel.cpp" />
     <ClCompile Include="Engine\Skinning\AnimationInstance.cpp" />
+    <ClCompile Include="Engine\Skinning\Bone.cpp" />
     <ClCompile Include="Engine\Skinning\Skeleton.cpp" />
     <ClCompile Include="Engine\Components\SpotLight.cpp" />
     <ClCompile Include="Engine\Systems\FPControllerSystem.cpp" />
